+++ src/main/java/org/gestern/gringotts/accountholder/PlayerAccountHolder.java
@@ -1,10 +1,11 @@
-package org.gestern.gringotts;
+package org.gestern.gringotts.accountholder;
 
 import java.util.HashMap;
 import java.util.Map;
 
 import org.bukkit.Bukkit;
 import org.bukkit.OfflinePlayer;
+import org.gestern.gringotts.AccountHolder;